Convert CustomerAccount to Customer and Add stripe
To create Customers in coinage, we need to construct all the relevant
stripe information as well. In order to do that I added stripe configs,
a gateway for wrapping stripe calls, and updated the customer model
to reflect stripe changes.

CustomerAccount was moved to Customer because the 'account' suffix did
not add anything to the description and was redundant (everything in
coinage was called 'account').