Transfer jonnykry/coinage to ghmeier/coinage
Since @ghmeier is now the primary maintainer of this repo, it's
being transferred for easy management.

Also, refactor helpers to actually use userID to determine roasterID
and then the stripe account. Before we were storing userID with
the account, but that would prevent roasters from being used by
multiple user accounts. Fixed it before it becomes a problem.